requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pymongo==4.7.3
langchain==0.2.1
langchain-core==0.2.3
langchain-groq==0.1.5
langgraph==0.1.1
streamlit==1.35.0
requests==2.32.3
werkzeug==3.0.3
pyjwt==2.8.0